ACD-4122: Removed Eventgen parameter to only check CIM app availability
