+++ pytest_splunk_addon/standard_lib/cim_tests/test_templates.py
@@ -4,6 +4,10 @@ Includes the test scenarios to check the CIM compatibility of an Add-on.
 """
 
 import logging
+import pytest
+
+INTERVAL = 3
+RETRIES = 3
 
 class CIMTestTemplates(object):
     """
@@ -18,4 +22,31 @@ class CIMTestTemplates(object):
     """
     logger = logging.getLogger("pytest-splunk-addon-cim-tests")
 
-    # TODO: Add Test scenarios 
+    @pytest.mark.parametrize(
+        "app_name", [pytest.param("Splunk_SA_CIM", marks=[pytest.mark.searchtime_cim])]
+    )
+    def test_app_installed(self, splunk_search_util, app_name, record_property):
+        """
+        This test case checks that addon is installed/enabled in the Splunk instance.
+
+        Args:
+            splunk_search_util (SearchUtil): Object that helps to search on Splunk.
+            app_name (string): Add-on name.
+            record_property (fixture): Document facts of test cases.
+        """
+
+        record_property("app_name", app_name)
+        # Search Query
+        search = "| rest /servicesNS/nobody/{}/configs/conf-app/ui".format(app_name)
+        self.logger.info(f"Executing the search query: {search}")
+        record_property("search", search)
+
+        result = splunk_search_util.checkQueryCountIsGreaterThanZero(
+            search, interval=INTERVAL, retries=RETRIES
+        )
+
+        assert result, (
+            f"\nMessage: App {app_name} is not installed/enabled in this Splunk instance."
+            f"The plugin requires the {app_name} to be installed/enabled in the Splunk instance."
+            f"Please install the app and execute the tests again."
+        )
